+++ src/short-urls/ShortUrlsFilteringBar.tsx
@@ -1,20 +1,21 @@
-import { faTag, faTags } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import type { OrderDir } from '@shlinkio/shlink-frontend-kit';
-import { Button, OrderingDropdown, SearchInput, Tooltip, useTooltip } from '@shlinkio/shlink-frontend-kit';
+import { OrderingDropdown, SearchInput } from '@shlinkio/shlink-frontend-kit';
+import type { TagsFilteringMode } from '@shlinkio/shlink-js-sdk/api-contract';
 import { clsx } from 'clsx';
 import type { FC } from 'react';
 import { useCallback, useState } from 'react';
 import type { FCWithDeps } from '../container/utils';
 import { componentFactory, useDependencies } from '../container/utils';
+import { DomainFilterDropdown } from '../domains/helpers/DomainFilterDropdown';
 import type { DomainsList } from '../domains/reducers/domainsList';
 import { useSetting } from '../settings';
-import type { TagsSelectorProps } from '../tags/helpers/TagsSelector';
+import type { TagsSearchDropdownProps } from '../tags/helpers/TagsSearchDropdown';
 import type { TagsList } from '../tags/reducers/tagsList';
 import { DateRangeSelector } from '../utils/dates/DateRangeSelector';
 import { formatIsoDate } from '../utils/dates/helpers/date';
 import type { DateInterval, DateRange } from '../utils/dates/helpers/dateIntervals';
 import { datesToDateRange } from '../utils/dates/helpers/dateIntervals';
+import { useFeature } from '../utils/features';
 import type { ShortUrlsOrder, ShortUrlsOrderableFields } from './data';
 import { SHORT_URLS_ORDERABLE_FIELDS } from './data';
 import type { ExportShortUrlsBtnProps } from './helpers/ExportShortUrlsBtn';
@@ -35,13 +36,13 @@ type ShortUrlsFilteringConnectProps = ShortUrlsFilteringBarProps & {
 
 type ShortUrlsFilteringBarDeps = {
   ExportShortUrlsBtn: FC<ExportShortUrlsBtnProps>;
-  TagsSelector: FC<TagsSelectorProps>;
+  TagsSearchDropdown: FC<TagsSearchDropdownProps>;
 };
 
 const ShortUrlsFilteringBar: FCWithDeps<ShortUrlsFilteringConnectProps, ShortUrlsFilteringBarDeps> = (
   { className, shortUrlsAmount, order, handleOrderBy, tagsList, domainsList },
 ) => {
-  const { ExportShortUrlsBtn, TagsSelector } = useDependencies(ShortUrlsFilteringBar);
+  const { ExportShortUrlsBtn, TagsSearchDropdown } = useDependencies(ShortUrlsFilteringBar);
   const [{
     search,
     tags,
@@ -54,6 +55,7 @@ const ShortUrlsFilteringBar: FCWithDeps<ShortUrlsFilteringConnectProps, ShortUrl
     tagsMode = 'any',
   }, toFirstPage] = useShortUrlsQuery();
   const visitsSettings = useSetting('visits');
+  const supportsFilterByDomain = useFeature('filterShortUrlsByDomain');
 
   const [activeInterval, setActiveInterval] = useState<DateInterval>();
   const setDates = useCallback(
@@ -70,69 +72,51 @@ const ShortUrlsFilteringBar: FCWithDeps<ShortUrlsFilteringConnectProps, ShortUrl
     (searchTerm: string) => toFirstPage({ search: !searchTerm ? undefined : searchTerm }),
     [toFirstPage],
   );
+  const setDomain = useCallback((domain?: string) => toFirstPage({ domain }), [toFirstPage]);
   const changeTagSelection = useCallback((newTags: string[]) => toFirstPage({ tags: newTags }), [toFirstPage]);
-  const toggleTagsMode = useCallback(
-    () => toFirstPage({ tagsMode: tagsMode === 'any' ? 'all' : 'any' }),
-    [tagsMode, toFirstPage],
-  );
-
-  const { anchor, tooltip } = useTooltip({ placement: 'left' });
+  const changeTagsMode = useCallback((tagsMode: TagsFilteringMode) => toFirstPage({ tagsMode }), [toFirstPage]);
 
   return (
     <div className={clsx('flex flex-col gap-y-4', className)}>
       <SearchInput defaultValue={search} onChange={setSearch} />
 
-      <div className="flex w-full">
-        <div className="flex-grow">
-          <TagsSelector
-            immutable
-            placeholder="With tags..."
-            tags={tagsList.tags}
-            selectedTags={tags}
-            onChange={changeTagSelection}
-            containerClassName={clsx(tags.length > 1 && '[&]:rounded-r-none')}
-          />
-        </div>
-        {tags.length > 1 && (
-          <>
-            <Button
-              variant="secondary"
-              onClick={toggleTagsMode}
-              aria-label="Change tags mode"
-              className="[&]:border-l-none [&]:rounded-l-none"
-              {...anchor}
-            >
-              <FontAwesomeIcon className="text-2xl" icon={tagsMode === 'all' ? faTags : faTag} />
-            </Button>
-            <Tooltip {...tooltip}>
-              {tagsMode === 'all' ? <>With <b>all</b> the tags</> : <>With <b>any</b> of the tags</>}
-            </Tooltip>
-          </>
-        )}
-      </div>
-
-      <div className="flex flex-col lg:flex-row-reverse gap-y-4">
-        <div className="lg:w-2/3 xl:w-1/2 inline-flex flex-col md:flex-row gap-x-2 gap-y-4">
-          <div className="grow">
-            <DateRangeSelector
-              defaultText="All short URLs"
-              dateRangeOrInterval={activeInterval ?? datesToDateRange(startDate, endDate)}
-              onDatesChange={setDates}
+      <div className="flex flex-col xl:flex-row-reverse justify-between gap-y-4">
+        <div className="min-w-2/3 inline-flex flex-col lg:flex-row gap-x-2 gap-y-4">
+          <div className="flex flex-col md:flex-row gap-x-2 gap-y-4 grow">
+            <div className="grow">
+              <DateRangeSelector
+                defaultText="All short URLs"
+                dateRangeOrInterval={activeInterval ?? datesToDateRange(startDate, endDate)}
+                onDatesChange={setDates}
+              />
+            </div>
+            <TagsSearchDropdown
+              title="Filter by tag"
+              prefix="With"
+              tags={tagsList.tags}
+              selectedTags={tags}
+              onTagsChange={changeTagSelection}
+              mode={tagsMode}
+              onModeChange={changeTagsMode}
+              buttonClassName="w-full"
             />
           </div>
-          <ShortUrlsFilterDropdown
-            selected={{
-              excludeBots: excludeBots ?? visitsSettings?.excludeBots,
-              excludeMaxVisitsReached,
-              excludePastValidUntil,
-              domain,
-            }}
-            onChange={toFirstPage}
-            domains={domainsList.loading ? undefined : domainsList.domains}
-          />
+          <div className={clsx('grid lg:flex gap-x-2 gap-y-4', { 'grid-cols-2': supportsFilterByDomain })}>
+            {supportsFilterByDomain && (
+              <DomainFilterDropdown domains={domainsList.domains} onChange={setDomain} value={domain} />
+            )}
+            <ShortUrlsFilterDropdown
+              selected={{
+                excludeBots: excludeBots ?? visitsSettings?.excludeBots,
+                excludeMaxVisitsReached,
+                excludePastValidUntil,
+              }}
+              onChange={toFirstPage}
+            />
+          </div>
         </div>
-        <div className="lg:w-1/3 xl:w-1/2 inline-flex gap-3">
-          <div className="max-lg:w-1/2 lg:hidden">
+        <div className="flex gap-2">
+          <div className="max-xl:w-1/2 xl:hidden">
             <OrderingDropdown
               containerClassName="[&]:block"
               buttonClassName="w-full"
@@ -142,7 +126,7 @@ const ShortUrlsFilteringBar: FCWithDeps<ShortUrlsFilteringConnectProps, ShortUrl
               onChange={({ field, dir }) => handleOrderBy(field, dir)}
             />
           </div>
-          <div className="max-lg:w-1/2">
+          <div className="max-xl:w-1/2">
             <ExportShortUrlsBtn amount={shortUrlsAmount} />
           </div>
         </div>
@@ -153,5 +137,5 @@ const ShortUrlsFilteringBar: FCWithDeps<ShortUrlsFilteringConnectProps, ShortUrl
 
 export const ShortUrlsFilteringBarFactory = componentFactory(
   ShortUrlsFilteringBar,
-  ['ExportShortUrlsBtn', 'TagsSelector'],
+  ['ExportShortUrlsBtn', 'TagsSearchDropdown'],
 );