+++ src/short-urls/data/index.ts
@@ -29,10 +29,3 @@ export type ExportableShortUrl = {
   tags: string;
   visits: number;
 };
-
-export type ShortUrlsFilter = {
-  excludeBots?: boolean;
-  excludeMaxVisitsReached?: boolean;
-  excludePastValidUntil?: boolean;
-  domain?: string;
-};