+++ src/short-urls/helpers/ShortUrlsFilterDropdown.tsx
@@ -1,24 +1,20 @@
 import { Dropdown } from '@shlinkio/shlink-frontend-kit';
 import { useCallback } from 'react';
-import type { Domain } from '../../domains/data';
-import { DEFAULT_DOMAIN } from '../../domains/data';
-import { useFeature } from '../../utils/features';
-import type { ShortUrlsFilter } from '../data';
 
-interface ShortUrlsFilterDropdownProps {
+export type ShortUrlsFilter = {
+  excludeBots?: boolean;
+  excludeMaxVisitsReached?: boolean;
+  excludePastValidUntil?: boolean;
+};
+
+export type ShortUrlsFilterDropdownProps = {
   onChange: (filters: ShortUrlsFilter) => void;
   selected?: ShortUrlsFilter;
+};
 
-  /**
-   * List of domains supported by the Shlink server.
-   * It is `undefined` while the domains are being loaded.
-   */
-  domains?: Domain[];
-}
-
-export const ShortUrlsFilterDropdown = ({ onChange, selected = {}, domains }: ShortUrlsFilterDropdownProps) => {
-  const supportsFilterByDomain = useFeature('filterShortUrlsByDomain');
-  const { excludeBots = false, excludeMaxVisitsReached = false, excludePastValidUntil = false, domain } = selected;
+export const ShortUrlsFilterDropdown = ({ onChange, selected = {} }: ShortUrlsFilterDropdownProps) => {
+  const { excludeBots = false, excludeMaxVisitsReached = false, excludePastValidUntil = false } = selected;
+  const amount = Number(excludeBots) + Number(excludeMaxVisitsReached) + Number(excludePastValidUntil);
 
   const partialUpdate = useCallback(
     (update: Partial<ShortUrlsFilter>) => onChange({ ...selected, ...update }),
@@ -30,7 +26,11 @@ export const ShortUrlsFilterDropdown = ({ onChange, selected = {}, domains }: Sh
   );
 
   return (
-    <Dropdown buttonContent="Filters" buttonClassName="w-full" menuAlignment="right">
+    <Dropdown
+      buttonContent={<span>More{amount > 0 ? <b> ({amount})</b> : ''}</span>}
+      buttonClassName="w-full"
+      menuAlignment="right"
+    >
       <Dropdown.Title>Visits:</Dropdown.Title>
       <Dropdown.Item selected={excludeBots} onClick={() => toggleFilter('excludeBots')}>
         Ignore visits from bots
@@ -45,40 +45,17 @@ export const ShortUrlsFilterDropdown = ({ onChange, selected = {}, domains }: Sh
         Exclude enabled in the past
       </Dropdown.Item>
 
-      {supportsFilterByDomain && (
-        <>
-          <Dropdown.Separator />
-          <Dropdown.Title>Domain: {!domains && <i>loading...</i>}</Dropdown.Title>
-          {domains?.map((d) => {
-            const value = d.isDefault ? DEFAULT_DOMAIN : d.domain;
-            const isSelected = domain === value;
-
-            return (
-              <Dropdown.Item
-                key={d.domain}
-                selected={isSelected}
-                onClick={() => partialUpdate({ domain: isSelected ? undefined : value })}
-              >
-                {d.domain}
-              </Dropdown.Item>
-            );
-          })}
-        </>
-      )}
-
       <Dropdown.Separator />
       <Dropdown.Item
         disabled={
           selected.excludeBots === undefined
           && selected.excludeMaxVisitsReached === undefined
           && selected.excludePastValidUntil === undefined
-          && selected.domain === undefined
         }
         onClick={() => onChange({
           excludeBots: undefined,
           excludeMaxVisitsReached: undefined,
           excludePastValidUntil: undefined,
-          domain: undefined,
         })}
         className="italic"
       >