+++ src/tags/helpers/TagsSelector.tsx
@@ -7,13 +7,13 @@ import type { ColorGenerator } from '../../utils/services/ColorGenerator';
 
 export type TagsSelectorProps = {
   onChange: (tags: string[]) => void;
-} & Pick<TagsAutocompleteProps, 'tags' | 'selectedTags' | 'placeholder' | 'immutable' | 'containerClassName'>;
+} & Pick<TagsAutocompleteProps, 'tags' | 'selectedTags' | 'placeholder'>;
 
 type TagsSelectorDeps = {
   ColorGenerator: ColorGenerator;
 };
 
-const TagsSelector: FCWithDeps<TagsSelectorProps, TagsSelectorDeps> = ({ onChange, placeholder, ...rest }) => {
+const TagsSelector: FCWithDeps<TagsSelectorProps, TagsSelectorDeps> = ({ onChange, ...rest }) => {
   const { ColorGenerator: colorGenerator } = useDependencies(TagsSelector);
   const shortUrlCreation = useSetting('shortUrlCreation');
   const searchMode = shortUrlCreation?.tagFilteringMode ?? 'startsWith';
@@ -24,7 +24,6 @@ const TagsSelector: FCWithDeps<TagsSelectorProps, TagsSelectorDeps> = ({ onChang
       onTagsChange={onChange}
       getColorForTag={(tag) => colorGenerator.getColorForKey(tag)}
       size="lg"
-      placeholder={placeholder ?? 'Add tags to the URL'}
       searchMode={searchMode}
     />
   );